Move libevent-specific dns code to libevent engine
